+++ src/shared/defaults.ts
@@ -3,6 +3,8 @@ export const DEFAULT_SETTINGS = {
   title: 'obs-tiles',
   currentConfigIndex: 0,
   selectConfigAtLaunch: false,
+  autoBackupConfigOnClose: false,
+  autoBackupConfigFolder: '',
   confirmBeforeStartStreaming: false,
   confirmBeforeStopStreaming: false,
   confirmBeforeStartRecording: false,